requirements.txt
google-cloud-storage
joblib
scikit-learn
flask